+++ pennylane/utils.py
@@ -25,6 +25,7 @@ from operator import matmul
 import warnings
 
 import numpy as np
+import scipy
 
 import pennylane as qml
 
@@ -103,6 +104,56 @@ def decompose_hamiltonian(H, hide_identity=False):
                 obs.append(functools.reduce(matmul, [t(i) for i, t in enumerate(term)]))
 
     return coeffs, obs
+
+
+def sparse_hamiltonian(H):
+    r"""Computes the sparse matrix representation a Hamiltonian in the computational basis.
+
+    Args:
+        H (~.Hamiltonian): Hamiltonian operator for which the matrix representation should be
+         computed
+
+    Returns:
+        coo_matrix: a sparse matrix in scipy coordinate list (COO) format with dimension
+        :math:`(2^n, 2^n)`, where :math:`n` is the number of wires
+
+    **Example:**
+
+    This function can be used by passing a `qml.Hamiltonian` object as:
+
+    >>> coeffs = [1, -0.45]
+    >>> obs = [qml.PauliZ(0) @ qml.PauliZ(1), qml.PauliY(0) @ qml.PauliZ(1)]
+    >>> H = qml.Hamiltonian(coeffs, obs)
+    >>> H_sparse = sparse_hamiltonian(H)
+    >>> H_sparse
+    <4x4 sparse matrix of type '<class 'numpy.complex128'>'
+        with 2 stored elements in COOrdinate format>
+
+    The resulting sparse matrix can be either used directly or transformed into a numpy array:
+
+    >>> H_sparse.toarray()
+    array([[ 1.+0.j  ,  0.+0.j  ,  0.+0.45j,  0.+0.j  ],
+           [ 0.+0.j  , -1.+0.j  ,  0.+0.j  ,  0.-0.45j],
+           [ 0.-0.45j,  0.+0.j  , -1.+0.j  ,  0.+0.j  ],
+           [ 0.+0.j  ,  0.+0.45j,  0.+0.j  ,  1.+0.j  ]])
+    """
+    if not isinstance(H, qml.Hamiltonian):
+        raise TypeError("Passed Hamiltonian must be of type `qml.Hamiltonian`")
+
+    n = len(H.wires)
+    matrix = scipy.sparse.coo_matrix((2 ** n, 2 ** n), dtype="complex128")
+
+    for coeffs, ops in zip(H.coeffs, H.ops):
+
+        obs = [scipy.sparse.coo_matrix(o.matrix) for o in qml.operation.Tensor(ops).obs]
+        mat = [scipy.sparse.eye(2, format="coo")] * n
+
+        for i, j in enumerate(ops.wires):
+            mat[j] = obs[i]
+
+        matrix += functools.reduce(lambda i, j: scipy.sparse.kron(i, j, format="coo"), mat) * coeffs
+
+    return matrix.tocoo()
 
 
 def _flatten(x):